fix types and styles

+++ cpp/open3d/core/hashmap/CPU/KvPairsCPU.hpp
@@ -51,9 +51,9 @@ public:
     std::atomic<int> heap_counter_; /* [1] */
 
 public:
-    int dsize_key_;
-    int dsize_value_;
-    int capacity_;
+    int64_t dsize_key_;
+    int64_t dsize_value_;
+    int64_t capacity_;
 
 public:
     // The value_ array's size is FIXED.
